move openai files
